requirements.txt
# Dependencias principales
lxml>=4.9.0  # Para procesamiento XML
python-dotenv>=0.19.0  # Para manejo de variables de entorno

# Desarrollo
typing-extensions>=4.0.0  # Para type hints en versiones antiguas de Python
pylint>=2.12.0  # Para análisis de código estático
black>=22.1.0  # Para formateo de código
pytest>=7.0.0  # Para pruebas unitarias

# Documentación
sphinx>=4.3.0  # Para documentación
sphinx-rtd-theme>=1.0.0  # Tema para la documentación
